requirements.txt
beautifulsoup4==4.14.3
Flask==3.1.3
numpy==2.4.3
Pillow==12.1.1
Requests==2.32.5
scikit_learn==1.8.0
skimage==0.0
